langchain-app/requirements.txt
langchain_cohere==0.1.8
langchain_community==0.2.6
langchain_core==0.2.10
langchain_openai==0.1.12
pinecone==4.0.0
python-dotenv==1.0.1
streamlit==1.36.0
